refactor: DB 연결 실패시 예외를 던지도록 수정
