+++ src/main/java/db/connection/DBConnection.java
@@ -1,4 +1,4 @@
-package db;
+package db.connection;
 
 import java.sql.Connection;
