+++ src/main/java/db/connection/MySQLConnection.java
@@ -1,33 +1,23 @@
-package db;
+package db.connection;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
 public class MySQLConnection implements DBConnection {
-    public static final MySQLConnection INSTANCE = new MySQLConnection();
-
     private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
     private static final String DATABASE = "chess"; // MySQL DATABASE 이름
     private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
     private static final String USERNAME = "user"; //  MySQL 서버 아이디
     private static final String PASSWORD = "password"; // MySQL 서버 비밀번호
 
-    private MySQLConnection() {
-    }
-
     @Override
     public Connection getConnection() {
         try {
             return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
         } catch (SQLException e) {
-            System.err.println("[ERROR] DB 연결 오류:" + e.getMessage());
-            e.printStackTrace();
-            return null;
-        }
-    }
+            throw new RuntimeException("[ERROR] DB에 연결에 실패했습니다.");
 
-    public static MySQLConnection getInstance() {
-        return INSTANCE;
+        }
     }
 }
